requirements.txt
numpy>=1.20.0
openai>=1.0.0
aiohttp>=3.8.0
pyyaml>=6.0
python-dotenv>=0.19.0
hvac>=2.1.0
transformers>=4.20.0
torch>=1.12.0
sentence-transformers>=2.2.0

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1